requirements.txt
aioredis==2.0.0
aiosmtplib==1.1.6
anyio==3.3.0
appdirs==1.4.4
asgiref==3.4.1
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
bcrypt==3.2.0
black==21.7b0
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
contextvars==2.4
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.5.2
fastapi==0.68.0
fastapi-mail==0.4.1
flake8==3.9.2
greenlet==1.1.1
h11==0.12.0
httpcore==0.13.6
httptools==0.2.0
httpx==0.18.2
idna==3.2
immutables==0.16
importlib-metadata==4.6.4
iniconfig==1.1.1
isort==5.9.3
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.0
passlib==1.7.4
pathspec==0.9.0
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
redis==3.5.3
regex==2021.8.3
requests==2.26.0
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.22
starlette==0.14.2
toml==0.10.2
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.14.0
watchgod==0.7
websockets==9.1
zipp==3.5.0
